Add fast iteration algorithm for numbers conundrums
This is a working draft of an exhaustive search algorithm. Number
of checked permutations and combinations is maybe larger than optimal
but at least there are non missing, as far as I can tell.

There are steps taken to winnow out solutions that are to similar,
but it still leaves plenty of the alike expressions. Filtering them
will be the next step.